+++ src/main/java/dao/PedidoDao.java
@@ -2,5 +2,5 @@ package dao;
 
 import model.Pedido;
 
-public interface PedidoDao extends GenericDao<Pedido, Integer> {
+public interface PedidoDao extends GenericDao<Pedido,Long> {
 }
