+++ src/main/java/dao/ProveedorDao.java
@@ -2,5 +2,5 @@ package dao;
 
 import model.Proveedor;
 
-public interface ProveedorDao  extends GenericDao<Proveedor, Integer> {
+public interface ProveedorDao  extends GenericDao<Proveedor, Long> {
 }
